Chore: Restructured Pages
