+++ src/Components/Pages/Homepage/Homepage.jsx
@@ -1,8 +1,8 @@
-import TopNav from "../Navbars/Top Navs/TopNav";
-import BottomNav from "../Navbars/Bottom Navs/BottomNav";
-import { HomepageLoadingCard } from "../Cards";
-import Feed from "../Feed/Feed";
-import CreatePost from "../Posts/CreatePost";
+import TopNav from "../../Navbars/Top Navs/TopNav";
+import BottomNav from "../../Navbars/Bottom Navs/BottomNav";
+import { HomepageLoadingCard } from "../../Cards";
+import Feed from "../../Feed/Feed";
+import CreatePost from "../../Posts/CreatePost";
 import styles from "./hompage.module.css";
 
 import { useSelector } from "react-redux";
