+++ src/Components/index.js
@@ -1,11 +1,11 @@
-import Homepage from "./Pages/Homepage";
+import Homepage from "./Pages/Homepage/Homepage";
 import SignIn from "./Pages/SignIn SignUp/SignIn";
 import SignUp from "./Pages/SignIn SignUp/SignUp";
 import NotFound from "./Pages/NotFound";
 import ProfilePage from "./Pages/Profile Page/ProfilePage";
-import NotificationPage from "./Pages/NotificationPage";
+import NotificationPage from "./Pages/Notification Page/NotificationPage";
 import SearchPage from "./Pages/Search Page/SearchPage";
-import PostPage from "./Pages/PostPage";
+import PostPage from "./Pages/Post Page/PostPage";
 import IndexPage from "./Pages/Index Page/IndexPage";
 
 export {
